Refactor schema tests
